Fix issue where some screenshots failed due to browser disconnect.
Some pages or provided HTML can cause the browser to crash / disconnect.
Those pages in my experience usually have a CORS error e.g. in their
font-face assets, or some sort of security warning. The exact cause is
unsure yet, but it's pretty random: in a specific html document, it
seems to fail anywhere between 1 in 30 or 1 in 90 times.

However with puppeteer-cluster we were not using its queuing system.
We need to wait synchronously and response to the express request, and
use puppeteer.execute(). With puppeteer-cluster's queue(), the request
could have been retried. This commit will add a retry (max. retries set
to 3, currently) to the execution of the puppeteer-cluster task itself.
Currently this seems to work and this page will get rendered.